Added JUnit Tests

+++ src/main/java/com/ToDoList2/ToDoList2/service/ToDoService.java
@@ -6,9 +6,9 @@ import com.ToDoList2.ToDoList2.entity.ToDo;
 
 public interface ToDoService {
     public List<ToDo> getAllToDos();
-    public ToDo getToDoById(Long id);
+    public ToDo getToDoById(Integer id);
     public ToDo createToDo(ToDo toDo);
-    public ToDo updateToDo(Long id, ToDo toDo);
-    public ToDo deleteToDo(Long id);
+    public ToDo updateToDo(Integer id, ToDo toDo);
+    public ToDo deleteToDo(Integer id);
     public void deleteAllToDos();
 }
